Simplify customer sync cron frequency values
