+++ Model/Config/Backend/Frequency.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Smaily\SmailyForMagento\Model\Config\Backend;
+
+use \Magento\Framework\Model\Context;
+use \Magento\Framework\Registry;
+use \Magento\Framework\App\Config\ScopeConfigInterface;
+use \Magento\Framework\App\Cache\TypeListInterface;
+use \Magento\Framework\App\Config\ValueFactory;
+use \Magento\Framework\Model\ResourceModel\AbstractResource;
+use \Magento\Framework\Data\Collection\AbstractDb;
+
+/**
+ * Adds frequecy from admin page to config db.
+ */
+class Frequency extends \Magento\Framework\App\Config\Value
+{
+    const CRON_STRING_PATH = 'crontab/default/jobs/smaily_subscriber_sync/schedule/cron_expr';
+
+    protected $configValueFactory;
+
+    public function __construct(
+        Context $context,
+        Registry $registry,
+        ScopeConfigInterface $config,
+        TypeListInterface $cacheTypeList,
+        ValueFactory $configValueFactory,
+        AbstractResource $resource = null,
+        AbstractDb $resourceCollection = null,
+        $runModelPath = '',
+        array $data = []
+    ) {
+        $this->configValueFactory = $configValueFactory;
+        parent::__construct($context, $registry, $config, $cacheTypeList, $resource, $resourceCollection, $data);
+    }
+
+    public function afterSave()
+    {
+        $frequency = $this->getData('groups/sync/fields/frequency/value');
+
+        try {
+            $this->configValueFactory->create()->load(
+                self::CRON_STRING_PATH,
+                'path'
+            )->setValue(
+                $frequency
+            )->setPath(
+                self::CRON_STRING_PATH
+            )->save();
+        } catch (\Exception $e) {
+            throw new \Exception(__('We can\'t save the cron expression.'));
+        }
+
+        return parent::afterSave();
+    }
+}
