+++ Model/Config/Source/Frequency.php
@@ -2,97 +2,34 @@
 
 namespace Smaily\SmailyForMagento\Model\Config\Source;
 
-class Frequency extends \Magento\Framework\App\Config\Value
+class Frequency implements \Magento\Framework\Option\ArrayInterface
 {
     /**
-     * Cron string path
-     */
-    const CRON_STRING_PATH = 'crontab/default/jobs/smaily_subscriber_sync/schedule/cron_expr';
-
-    /**
-     * Cron model path
-     */
-    const CRON_MODEL_PATH = 'crontab/default/jobs/smaily_subscriber_sync/run/model';
-
-    /**
-     * @var \Magento\Framework\App\Config\ValueFactory
-     */
-    protected $configValueFactory;
-
-    /**
-     * @var string
-     */
-    protected $runModelPath = '';
-
-    /**
-     * @param \Magento\Framework\Model\Context $context
-     * @param \Magento\Framework\Registry $registry
-     * @param \Magento\Framework\App\Config\ScopeConfigInterface $config
-     * @param \Magento\Framework\App\Cache\TypeListInterface $cacheTypeList
-     * @param \Magento\Framework\App\Config\ValueFactory $configValueFactory
-     * @param \Magento\Framework\Model\ResourceModel\AbstractResource $resource
-     * @param \Magento\Framework\Data\Collection\AbstractDb $resourceCollection
-     * @param string $runModelPath
-     * @param array $data
-     */
-    public function __construct(
-        \Magento\Framework\Model\Context $context,
-        \Magento\Framework\Registry $registry,
-        \Magento\Framework\App\Config\ScopeConfigInterface $config,
-        \Magento\Framework\App\Cache\TypeListInterface $cacheTypeList,
-        \Magento\Framework\App\Config\ValueFactory $configValueFactory,
-        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
-        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
-        $runModelPath = '',
-        array $data = []
-    ) {
-        $this->runModelPath = $runModelPath;
-        $this->configValueFactory = $configValueFactory;
-        parent::__construct($context, $registry, $config, $cacheTypeList, $resource, $resourceCollection, $data);
-    }
-
-    /**
-     * {@inheritdoc}
+     * Get Option values for customer sync frequency fields.
      *
-     * @return $this
-     * @throws \Exception
+     * @return array
      */
-    public function afterSave()
+    public function toOptionArray()
     {
-        $time = $this->getData('groups/sync/fields/time/value');
-        $frequency = $this->getData('groups/sync/fields/frequency/value');
-
-        $cronExprArray = [
-            intval($time[1]), //Minute
-            intval($time[0]), //Hour
-            $frequency == \Magento\Cron\Model\Config\Source\Frequency::CRON_MONTHLY ? '1' : '*', //Day of the Month
-            '*', //Month of the Year
-            $frequency == \Magento\Cron\Model\Config\Source\Frequency::CRON_WEEKLY ? '1' : '*', //Day of the Week
+        $list = [
+            [
+                'value' => '0 */4 * * *',
+                'label' => 'Every 4 hours',
+            ],
+            [
+                'value' => '0 */12 * * *',
+                'label' => 'Twice a day',
+            ],
+            [
+                'value' => '0 * */1 * *',
+                'label' => 'Every day',
+            ],
+            [
+                'value' => '0 0 * * 0',
+                'label' => 'Once a week',
+            ],
         ];
 
-        $cronExprString = join(' ', $cronExprArray);
-
-        try {
-            $this->configValueFactory->create()->load(
-                self::CRON_STRING_PATH,
-                'path'
-            )->setValue(
-                $cronExprString
-            )->setPath(
-                self::CRON_STRING_PATH
-            )->save();
-            $this->configValueFactory->create()->load(
-                self::CRON_MODEL_PATH,
-                'path'
-            )->setValue(
-                $this->runModelPath
-            )->setPath(
-                self::CRON_MODEL_PATH
-            )->save();
-        } catch (\Exception $e) {
-            throw new \Exception(__('We can\'t save the cron expression.'));
-        }
-
-        return parent::afterSave();
+        return $list;
     }
 }
