+++ etc/adminhtml/system.xml
@@ -52,11 +52,8 @@
                 </field>
                 <field id="frequency" translate="label" type="select" sortOrder="90" showInDefault="1" showInWebsite="0" showInStore="0">
                     <label>Frequency</label>
-                    <source_model>Magento\Cron\Model\Config\Source\Frequency</source_model>
-                    <backend_model>Smaily\SmailyForMagento\Model\Config\Source\Frequency</backend_model>
-                </field>
-                <field id="time" translate="label" type="time" sortOrder="100" showInDefault="1" showInWebsite="0" showInStore="0">
-                    <label>Start Time</label>
+                    <source_model>Smaily\SmailyForMagento\Model\Config\Source\Frequency</source_model>
+                    <backend_model>Smaily\SmailyForMagento\Model\Config\Backend\Frequency</backend_model>
                 </field>
             </group>
             <group id="abandoned" translate="label" type="text" sortOrder="40" showInDefault="1" showInWebsite="0" showInStore="0">
